+++ packages/delta-protocol-common/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 0.7.0
+## 0.7.2
+
+* Use the `propertyValueSerializerWith` function from the `class-core` package for the `deltaToCommandTranslator` and `deltaToEventTranslator` functions.
+
+
+## 0.7.1
+
+(The 0.7.0 release was deprecated because its `validation` package was faulty.)
 
 Initial creation and publication of this package.